music_spider/requirements.txt
pymongo
twisted
redis
lxml
beautifulsoup4
pysocks
pycrypto
requests>=2.11.1
msgpack-python